etl/streaming/requirements.txt
kafka-python-ng
websocket-client==1.6.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis>=5.0.0
